+++ src/Reservation.java
@@ -155,7 +155,7 @@ public class Reservation {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(double price) {
         this.price = price;
     }
